# Avoiding bound checks in setunion code