--- a/setutil_generic.go
+++ b/setutil_generic.go
@@ -21,8 +21,8 @@ func union2by2(set1 []uint16, set2 []uint16, buffer []uint16) int {
 	}
 	s1 := set1[k1]
 	s2 := set2[k2]
-	buffer = buffer[:cap(buffer)]
-	for {
+	buffer = buffer[:len1 + len2]
+	for pos < uint(len(buffer)) {
 		if pos >= uint(len(buffer)) {
 			panic("You provided a buffer with insufficient capacity")
 		}
@@ -35,9 +35,7 @@ func union2by2(set1 []uint16, set2 []uint16, buffer []uint16) int {
 				pos += len2 - k2
 				return int(pos)
 			}
-			if k1 < len1 {
-			  s1 = set1[k1]
-			}
+			s1 = set1[k1]
 		} else if s1 == s2 {
 			buffer[pos] = s1
 			pos++