@puzpuzpuz 2021-07-12
For me, Go v1.16.5's produces `CMPQ` + `JBE` instructions for bounds checks in slices. The branch should never be taken in the correct code, so it seems to me that the check should be very cheap.

I've tried to make `union2by2`'s loop fully branchless, except for the loop condition here:
https://github.com/puzpuzpuz/roaring/commit/56306171f5fbe8726d082c4a9dd02538933d252f

The results are worse than the branchy implementation, so I'm either doing something wrong (which is quite probable) or due to some factors (data dependencies between loop iterations? cost of additional arithmetical operations? something else) it's just not worth it. I'm under an impression that the proper way to approach `union2by2` is to use SIMD instructions just like it's done in the C++ roaring bitmap library, but obviously such implementation in Go assembler would be hard to maintain.