@lemire 2021-07-13
@puzpuzpuz

It is damn hard to benchmark branchy code. For example, to see the benefits of a branchless approach, you have to have mispredicted branches to begin with, but most benchmarking techniques minimize mispredicted branches by looping over the same data. 

We tend to always underestimate the value of branchless approaches.

> The branch should never be taken in the correct code, so it seems to me that the check should be very cheap.

Very cheap is relative. You still incur the two instructions (maybe less with fusion) per access. A union function is basically just that... access, compare, access, compare... so if access cost you even a bit more, it should show up in the performance.

But you are, of course, correct that the real solution is to use assembly.

Your branchless approach is interesting by, given that this is Go, it may not compile to efficient assembly. Have you looked at the assembly?