@puzpuzpuz 2021-07-13
> For example, to see the benefits of a branchless approach, you have to have mispredicted branches to begin with, but most benchmarking techniques minimize mispredicted branches by looping over the same data.

Yeah, I'm aware of the problem. That's why I'm using a large array of randomly generated arrays (~1M of numbers in total) here: https://github.com/puzpuzpuz/roaring/commit/56306171f5fbe8726d082c4a9dd02538933d252f#diff-c2898d6069ac5db5285b63cc9066e814584908dd34801fde55919838e1ef53bbR176

In theory, 1M of iterations should be enough to get rid of the effects of the branch predictor. For instance, if I change the benchmark to the similar [approach](https://github.com/RoaringBitmap/roaring/pull/239/files#diff-c2898d6069ac5db5285b63cc9066e814584908dd34801fde55919838e1ef53bbR246) as you did in the shortgun intersection experiment (generate random array on each iteration), `perf` linux utility reports me 2.96% of all branches for the `union2by2` scenario (the original branchy implementation), but with the approach from my branch it reports 5.76% of all branches. That seems to be an indication of the fact that the second approach eliminates effects of the branch prediction in a more efficient manner to me, but I may be wrong.

> Have you looked at the assembly?

Yes, I did. I'm not sure about how good the compiler output is in terms of the "cost" of the emitted instructions, but I didn't notice any branches except for the loop condition one. Most likely those instructions are rather expensive since I see higher instruction per cycle value reported by `perf` for the `union2by2_branchless` function (2.23 vs 1.46 insn per cycle).