@lemire 2021-07-13
> In theory, 1M of iterations should be enough to get rid of the effects of the branch predictor. 

I agree that what you did looks good. 

Ideally, you'd want the intersection size to vary unpredictably between 'all' to 'nothing'... I don't expect your approach to be good at that... but that's not likely to be an issue in this instance.

>  I see higher instruction per cycle value

We want more instructions per cycle. Being at 1.46 ins per cycle is *bad*. Keep in mind that recent processors like the Apple M1 can consistently retire 8 instructions per cycle.

But it does explain why trying to get a leaner code does not help. If we are at 1.46 ins per cycle, then trying to save instructions will not help us because we have plenty of room for more instructions. This suggest that we are primarily limited by data dependencies issues. And bound checks do nothing to harm data dependency.

This suggests that the move forward would be algorithmic in nature, at least for big superscalar processors.


Note that we have a faster assembly version of this function for ARM:

https://github.com/RoaringBitmap/roaring/blob/master/setutil_arm64.s

It is not algorithmic in nature. But it is possible that the gains are explained by the fact that it is tested on relatively small cores.