@puzpuzpuz 2021-07-13
Thanks for you inputs!

> We want more instructions per cycle. Being at 1.46 ins per cycle is _bad_. Keep in mind that recent processors like the Apple M1 can consistently retire 8 instructions per cycle.

I misinterpreted the metric, my bad (probably, due to incredibly hot weather I have locally). The higher insn per cycle measurement we get, the better since in general it means that the code benefits from ILP.

Thanks for the correction!

> This suggest that we are primarily limited by data dependencies issues. And bound checks do nothing to harm data dependency.

I'm also under the same impression so far.

> This suggests that the move forward would be algorithmic in nature, at least for big superscalar processors.

Makes sense. As for the ARM `union2by2` version, it's probably faster than the generic one due to Go compiler producing suboptimal code for arm64. But I'm just guessing since I don't have an ARM machine around to experiment with it.