@lemire 2021-07-13
> But I'm just guessing since I don't have an ARM machine around to experiment with it.

It may depend on the specific ARM core microarchitecture. There are many small ARM cores that struggle to retire 2 instructions per cycle. But there are also powerful ARM cores.

Historically, x64 cores have been more similar (large superscalar cores)... there are tiny x86 cores used for mobile devices but most of us don't use them very much.

On small cores, saving a few instructions can make a big difference.

Anyhow, I think that the long-term fix might be algorithmic.

(Note: your investigation has been fruitful. Let us consider algorithmic improvements!)