@puzpuzpuz 2021-07-14
A minor update on the branchless implementation. I was able to simplify it and got rid of the lookup tables. The main part of the changes in the loop code may be seen here: https://github.com/puzpuzpuz/roaring/commit/a2b25106daaaa969d33ed4c917d284a1c7e10cc1#diff-7916b4e45da125122d637e8abb339ad262f2efc9fde0e2ceb6671ae9e75fd298R36

With this update, the `union2by2_branchless` is now slightly faster than the baseline branchy function in the microbenchmark.

Before:
```bash
BenchmarkUnion2by2/union2by2_branchless-8         	      40	  29035424 ns/op
BenchmarkUnion2by2/union2by2-8                    	      48	  24743344 ns/op
```

After:
```bash
BenchmarkUnion2by2/union2by2_branchless-8         	      50	  22842745 ns/op
BenchmarkUnion2by2/union2by2-8                    	      46	  24849983 ns/op
```

I'm glad to finally see the improvement. On the other hand, even in such synthetic scenario it's not significant enough to be considered as a valuable optimization.