@lemire 2021-07-14
@puzpuzpuz A consistent 5% speed boost would be valuable.