@puzpuzpuz 2021-07-14
> @puzpuzpuz A consistent 5% speed boost would be valuable.

OK, let me run real-roaring-datasets and synthetic benchmarks and see if the branchless function makes any difference. If it does, I'll submit a PR. Thanks for your inputs once again.