@lemire 2021-07-14
What happens when the two values are equal in your algorithm?